prefect-dags/requirements.txt
polars-lts-cpu
pandas
skimpy
matplotlib
numpy
pandarallel
plotly
fuzzywuzzy
pickle
prefect==2.16.2
seaborn
surprise
python-box